Restructure text version of P4_16 allowed constructs a bit
There is no actual change in substantive content represented by this change.

action declarations are allowed at the top level, and inside control
blocks, and this is represented before and after this change.  This
change simply gives the full details for what can occur inside of an
action in the place where it is perhaps more commonly found, inside of
a control block.  The top level occurrence of action in the text
refers to the later one for details.