Add extract and emit of TCP options padding to tcp-options-parser.p4
Also add emit of that padding to tcp-options-parser2.p4.  I neglected
to add that in the previous commit.

Also removed optional varbit<> stuff in tcp-options-parser.p4.  Its
presence was not the reason that the program is not compiling using
p4c-bm2-ss.  It is because of the presence of header_union.