Change bash scripts for updating/building p4c to not pass command line args to cmake
Doing so meant that the command 'build-p4c.sh update' was broken.

Removing this capability from both scripts makes a bit of sense for
consistency's sake, vs. only removing it from one.  I may remove
update-p4c.sh entirely in the future, since it is a subset of the
functionality of build-p4c.sh