Minor tweaks to demo1, demo2, demo3 programs
There is no change in behavior, and I have verified that the README
instructions for creating table entries still work with these changes,
and in fact that the BMv2 JSON files generated are nearly identical.

I merely wanted to make some of the struct names and control/parser
names more like the v1model P4_16 template code in the templates
directory.